Optimization

+++ Application/Commands/Entities/Books/DeleteBookCommand.cs
@@ -6,8 +6,16 @@ public class DeleteBookCommand(IUnitOfWork unitOfWork)
     public async Task<BasicCreateDeleteResponse> Execute(DeleteBookRequest request, CancellationToken cancellationToken)
     {
         var bookRepo = unitOfWork.GetRepository<Book>();
-        var books = await bookRepo.Get(x => x.Id.Value == request.Id.Value, cancellationToken);
+
+        // Загружаем книгу вместе со связанными RoomBook
+        var books = await bookRepo.GetWithIncludesAsync(
+            b => b.Id.Value == request.Id.Value,
+            IncludePaths.Book.RoomBook);
         var book = books.FirstOrDefault() ?? throw new BookNotFoundException();
+
+        // Проверяем, есть ли у книги привязки к комнатам
+        if (book.RoomBook != null && book.RoomBook.Any())
+            throw new LibValidationException { ExceptionDetails = ["Невозможно удалить книгу, так как она присутствует в одной или нескольких комнатах"] };
 
         await bookRepo.Remove(book, cancellationToken);
         await unitOfWork.SaveChangesAsync(cancellationToken);
